+++ exporter/otlpexporter/factory_test.go
@@ -16,6 +16,7 @@ package otlpexporter
 
 import (
 	"context"
+	"path"
 	"testing"
 	"time"
 
@@ -167,7 +168,7 @@ func TestCreateTracesExporter(t *testing.T) {
 					Endpoint: endpoint,
 					TLSSetting: configtls.TLSClientSetting{
 						TLSSetting: configtls.TLSSetting{
-							CAFile: "testdata/test_cert.pem",
+							CAFile: path.Join("testdata", "test_cert.pem"),
 						},
 					},
 				},