+++ exporter/otlpexporter/otlp_test.go
@@ -17,6 +17,7 @@ package otlpexporter
 import (
 	"context"
 	"net"
+	"path"
 	"path/filepath"
 	"runtime"
 	"sync"
@@ -81,8 +82,8 @@ func otlpTracesReceiverOnGRPCServer(ln net.Listener, useTLS bool) (*mockTracesRe
 	if useTLS {
 		_, currentFile, _, _ := runtime.Caller(0)
 		basepath := filepath.Dir(currentFile)
-		certpath := filepath.Join(basepath, "testdata/test_cert.pem")
-		keypath := filepath.Join(basepath, "testdata/test_key.pem")
+		certpath := filepath.Join(basepath, path.Join("testdata", "test_cert.pem"))
+		keypath := filepath.Join(basepath, path.Join("testdata", "test_key.pem"))
 
 		creds, err := credentials.NewServerTLSFromFile(certpath, keypath)
 		if err != nil {
@@ -203,6 +204,8 @@ func TestSendTraces(t *testing.T) {
 		},
 	}
 	set := componenttest.NewNopExporterCreateSettings()
+	set.BuildInfo.Description = "Collector"
+	set.BuildInfo.Version = "1.2.3test"
 	exp, err := factory.CreateTracesExporter(context.Background(), set, cfg)
 	require.NoError(t, err)
 	require.NotNil(t, exp)
@@ -247,7 +250,10 @@ func TestSendTraces(t *testing.T) {
 	assert.EqualValues(t, 2, atomic.LoadInt32(&rcv.requestCount))
 	assert.EqualValues(t, td, rcv.GetLastRequest())
 
-	require.EqualValues(t, rcv.GetMetadata().Get("header"), expectedHeader)
+	md := rcv.GetMetadata()
+	require.EqualValues(t, md.Get("header"), expectedHeader)
+	require.Equal(t, len(md.Get("User-Agent")), 1)
+	require.Contains(t, md.Get("User-Agent")[0], "Collector/1.2.3test")
 }
 
 func TestSendTracesWhenEndpointHasHttpScheme(t *testing.T) {
@@ -344,6 +350,8 @@ func TestSendMetrics(t *testing.T) {
 		},
 	}
 	set := componenttest.NewNopExporterCreateSettings()
+	set.BuildInfo.Description = "Collector"
+	set.BuildInfo.Version = "1.2.3test"
 	exp, err := factory.CreateMetricsExporter(context.Background(), set, cfg)
 	require.NoError(t, err)
 	require.NotNil(t, exp)
@@ -388,7 +396,10 @@ func TestSendMetrics(t *testing.T) {
 	assert.EqualValues(t, 4, atomic.LoadInt32(&rcv.totalItems))
 	assert.EqualValues(t, md, rcv.GetLastRequest())
 
-	require.EqualValues(t, rcv.GetMetadata().Get("header"), expectedHeader)
+	mdata := rcv.GetMetadata()
+	require.EqualValues(t, mdata.Get("header"), expectedHeader)
+	require.Equal(t, len(mdata.Get("User-Agent")), 1)
+	require.Contains(t, mdata.Get("User-Agent")[0], "Collector/1.2.3test")
 }
 
 func TestSendTraceDataServerDownAndUp(t *testing.T) {
@@ -545,6 +556,8 @@ func TestSendLogData(t *testing.T) {
 		},
 	}
 	set := componenttest.NewNopExporterCreateSettings()
+	set.BuildInfo.Description = "Collector"
+	set.BuildInfo.Version = "1.2.3test"
 	exp, err := factory.CreateLogsExporter(context.Background(), set, cfg)
 	require.NoError(t, err)
 	require.NotNil(t, exp)
@@ -586,4 +599,8 @@ func TestSendLogData(t *testing.T) {
 	assert.EqualValues(t, 2, atomic.LoadInt32(&rcv.requestCount))
 	assert.EqualValues(t, 2, atomic.LoadInt32(&rcv.totalItems))
 	assert.EqualValues(t, ld, rcv.GetLastRequest())
+
+	md := rcv.GetMetadata()
+	require.Equal(t, len(md.Get("User-Agent")), 1)
+	require.Contains(t, md.Get("User-Agent")[0], "Collector/1.2.3test")
 }