+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 🛑 Breaking changes 🛑
 
 - Remove `config.NewConfigMapFrom[File|Buffer]`, add testonly version (#4502)
+- `configtls`: TLS 1.2 is the new default mininum version (#4503)
 
 ## 💡 Enhancements 💡
 
@@ -28,6 +29,7 @@
 
 - Add semconv 1.7.0 and 1.8.0 (#4452)
 - Added `feature-gates` CLI flag for controlling feature gate state. (#4368)
+- Add a default user-agent header to the OTLP/gRPC and OTLP/HTTP exporters containing collector build information (#3970)
 
 ## v0.39.0 Beta